==> 5_Normalize_Audio_Data.ipynb <==
# Imports

import numpy as np
import progressbar as pb

# Configuration

# directory to store intermediate and final results of the experiment
EXPERIMENT_DIR      = "/home/schindlera/experiments/ismir2020_reviews/"

AUDIO_FILENAME_STEM = "melspec_128_10seconds_2ch"

# Normalize Data

## Calculate Means from Train Partition

Calculate means over time for each Mel-band and channel

audio_filename_train = "%s/%s_train.npz" % (EXPERIMENT_DIR, AUDIO_FILENAME_STEM)

with np.load(audio_filename_train, allow_pickle=True) as npz:
    data     = npz["data"]
    trackids = npz["track_ids"]

If you have already stored the values, use the code below to load them

normalization_param_filename = "%s/normalization_parameters.npz" % (EXPERIMENT_DIR)

with np.load(normalization_param_filename, allow_pickle=True) as npz:
    means = npz["means"]
    stds  = npz["stds"]

means = data.mean(axis=(0,2), keepdims=True)
stds     = data.std(axis=(0,2), keepdims=True)

normalization_param_filename = "%s/normalization_parameters.npz" % (EXPERIMENT_DIR)

np.savez(normalization_param_filename, means=means, stds=stds)

## Normalize all Experiment Partitions

First, normalize the train partition, because the data is already loaded in memory

for i in pb.ProgressBar()(np.arange(0, data.shape[0], 100)):
    
    start = i
    stop  = start + 100
    
    data[start:stop,:,:,:] -= means
    data[start:stop,:,:,:] /= stds

Store normalized audio data

audio_filename_train = "%s/%s_norm_train.npz" % (EXPERIMENT_DIR, AUDIO_FILENAME_STEM)

np.savez(audio_filename_train, 
         data      = data, 
         track_ids = trackids)

del data, trackids

for par in ["val", "test"]:

    # load audio data
    audio_filename_par = "%s/%s_%s.npz" % (EXPERIMENT_DIR, AUDIO_FILENAME_STEM, par)
    
    with np.load(audio_filename_par, allow_pickle=True) as npz:
        data     = npz["data"]
        trackids = npz["track_ids"]
        
    # normalize audio data
    for i in pb.ProgressBar()(np.arange(0, data.shape[0], 100)):

        start = i
        stop  = start + 100

        data[start:stop,:,:,:] -= means
        data[start:stop,:,:,:] /= stds
        

    # store normalized audio data
    audio_filename_train = "%s/%s_norm_%s.npz" % (EXPERIMENT_DIR, AUDIO_FILENAME_STEM, par)

    np.savez(audio_filename_train, 
             data      = data, 
             track_ids = trackids)
        
        
    del data, trackids
